pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "schema-sentinel"
version = "2.1.0"
description = "Compare database metadata across environments - built for Snowflake data engineers"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "Igor Gladyshev", email = "92756539+Igladyshev@users.noreply.github.com"},
]
maintainers = [
    {name = "Igor Gladyshev", email = "92756539+Igladyshev@users.noreply.github.com"},
]
keywords = [
    "sql",
    "database",
    "metadata",
    "comparison",
    "snowflake",
    "schema",
    "data-engineering",
    "devops",
    "database-schema",
    "schema-migration",
    "data-quality",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]

dependencies = [
    "click>=8.1.3",
    "pandas>=2.0.1",
    "snowflake-connector-python>=3.0.3",
    "cryptography>=39.0.2",
    "snowflake-sqlalchemy>=1.4.7",
    "jinja2>=3.1.2",
    "tinyhtml>=1.2.0",
    "markdownmaker>=0.4.0",
    "tabulate>=0.9.0",
    "simple-ddl-parser>=0.30.0",
    "snakemd>=2.1.0",
    "sql-formatter>=0.6.2",
    "sqlparse>=0.4.4",
    "pdfkit>=1.0.0",
    "alembic[tz]>=1.11.3",
    "SQLAlchemy>=1.4.49,<2.0",
    "typer>=0.9.0",
    "sqlalchemy-json>=0.7.0",
]

[project.urls]
Homepage = "https://github.com/Igladyshev/schema-sentinel"
Documentation = "https://github.com/Igladyshev/schema-sentinel#readme"
Repository = "https://github.com/Igladyshev/schema-sentinel"
Issues = "https://github.com/Igladyshev/schema-sentinel/issues"
Changelog = "https://github.com/Igladyshev/schema-sentinel/blob/master/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pre-commit>=3.3.3",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pdoc>=14.0.0",
]
jupyter = [
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "ipykernel>=6.0.0",
]

[project.scripts]
schema-sentinel = "schema_sentinel.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["schema_sentinel"]

[tool.ruff]
line-length = 120
target-version = "py39"
extend-exclude = ["*.ipynb"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=schema_sentinel --cov-report=term-missing"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
